Add files via upload
The main design  of this project is  to predict  the survival of passengers based off a set of data. i had predict titanic survivors using logistic regression , RandomForestClassifier and Decision tree Algorithms.The most important factors when determining the chance of survival using Survival rate ,Fare and Gender.The majority of survivors where from small families and Young people had the greatest chance of survival.